requirements.txt
pyodbc===4.0.34
streamlit
python-dotenv
openai
promptflow
